KOGITO-2299: Refactor Domain Explorer related components
