+++ packages/common/index.ts
@@ -1,31 +1,16 @@
-export {
-  default as KogitoSpinner
-} from './src/components/Atoms/KogitoSpinner/KogitoSpinner';
-export {
-  default as PageToolbar
-} from './src/components/Molecules/PageToolbar/PageToolbar';
-export {
-  default as KogitoPageLayout
-} from './src/components/Templates/KogitoPageLayout/KogitoPageLayout';
-export {
-  default as AboutModalBox
-} from './src/components/Molecules/AboutModalBox/AboutModalBox';
-export {
-  default as BrandLogo
-} from './src/components/Atoms/BrandLogo/BrandLogo';
-export {
-  default as PageNotFound
-} from './src/components/Molecules/PageNotFound/PageNotFound';
-export {
-  default as ServerUnavailable
-} from './src/components/Molecules/ServerUnavailable/ServerUnavailable';
+export { default as KogitoSpinner } from './src/components/Atoms/KogitoSpinner/KogitoSpinner';
+export { default as PageToolbar } from './src/components/Molecules/PageToolbar/PageToolbar';
+export { default as KogitoPageLayout } from './src/components/Templates/KogitoPageLayout/KogitoPageLayout';
+export { default as AboutModalBox } from './src/components/Molecules/AboutModalBox/AboutModalBox';
+export { default as BrandLogo } from './src/components/Atoms/BrandLogo/BrandLogo';
+export { default as PageNotFound } from './src/components/Molecules/PageNotFound/PageNotFound';
+export { default as ServerUnavailable } from './src/components/Molecules/ServerUnavailable/ServerUnavailable';
 export { default as NoData } from './src/components/Molecules/NoData/NoData';
-export {
-  default as ServerErrors
-} from './src/components/Molecules/ServerErrors/ServerErrors';
-export {
-  default as ProcessDescriptor
-} from './src/components/Molecules/ProcessDescriptor/ProcessDescriptor';
+export { default as ServerErrors } from './src/components/Molecules/ServerErrors/ServerErrors';
+export { default as ProcessDescriptor } from './src/components/Molecules/ProcessDescriptor/ProcessDescriptor';
 export * from './src/components/Atoms/KogitoEmptyState/KogitoEmptyState';
+export { default as LoadMore } from './src/components/Atoms/LoadMore/LoadMore';
+export { default as DomainExplorer } from './src/components/Organisms/DomainExplorer/DomainExplorer';
+export { default as DomainExplorerListDomains } from './src/components/Organisms/DomainExplorerListDomains/DomainExplorerListDomains';
 export * from './src/utils/OuiaUtils';
 export * from './src/graphql/types';
