+++ packages/common/src/components/Molecules/ServerErrors/ServerErrors.css
@@ -1,4 +1,4 @@
-.kogito-management-console--Server-Errors__text-color {
+.kogito-common--Server-Errors__text-color {
     color: var(--pf-global--primary-color--100);
     cursor: pointer
 }
