+++ packages/common/src/components/Molecules/ServerUnavailable/ServerUnavailable.css
@@ -1,3 +1,3 @@
-kogito-common--PageLayout {
+.kogito-common--PageLayout {
     --pf-c-page__header-brand-link--c-brand--MaxHeight: 42px;
 }
