+++ packages/common/src/components/Organisms/DomainExplorer/DomainExplorer.css
@@ -1,0 +1,3 @@
+.kogito-common--domain-explorer__table-OverFlow {
+    overflow-y: auto;
+}
