+++ packages/management-console/src/components/Organisms/ProcessListTable/__mocks__/ProcessListTable.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-const DataListComponent = (): React.ReactElement => {
+const ProcessListComponent = (): React.ReactElement => {
   return <></>;
 };
 
-export default DataListComponent;
+export default ProcessListComponent;
