+++ packages/management-console/src/components/Templates/DomainExplorerLandingPage/tests/DomainExplorerLandingPage.test.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import DomainExplorerLandingPage from '../DomainExplorerLandingPage';
-import { GraphQL } from '@kogito-apps/common';
 import { MemoryRouter as Router } from 'react-router-dom';
+import { MockedProvider } from '@apollo/react-testing';
 
-jest.mock('@kogito-apps/common/src/graphql/types');
 describe('Domain Explorer Landing Page Component', () => {
   const props = {
     ouiaContext: {
@@ -13,93 +12,13 @@ describe('Domain Explorer Landing Page Component', () => {
     } as any
   };
   it('Snapshot test', () => {
-    // @ts-ignore
-    GraphQL.useGetQueryFieldsQuery.mockReturnValue({
-      loading: false,
-      data: {
-        __type: {
-          fields: [
-            {
-              name: 'Jobs',
-              args: []
-            },
-            {
-              name: 'Travels',
-              args: []
-            },
-            {
-              name: 'visaApplication',
-              args: []
-            }
-          ]
-        }
-      }
-    });
     const wrapper = mount(
-      <Router keyLength={0}>
-        <DomainExplorerLandingPage {...props} />
-      </Router>
+      <MockedProvider mocks={[]} addTypename={false}>
+        <Router keyLength={0}>
+          <DomainExplorerLandingPage {...props} />
+        </Router>
+      </MockedProvider>
     );
-    expect(GraphQL.useGetQueryFieldsQuery).toHaveBeenCalled();
-    expect(wrapper).toMatchSnapshot();
-  });
-  it('Assertions', () => {
-    // @ts-ignore
-    GraphQL.useGetQueryFieldsQuery.mockReturnValue({
-      loading: false,
-      data: {
-        __type: {
-          fields: [
-            {
-              name: 'Travels',
-              args: [
-                {
-                  name: 'where',
-                  type: {
-                    kind: 'INPUT_OBJECT',
-                    name: 'TravelsArgument'
-                  }
-                },
-                {
-                  name: 'orderBy',
-                  type: {
-                    kind: 'INPUT_OBJECT',
-                    name: 'TravelsOrderBy'
-                  }
-                }
-              ],
-              type: { ofType: { name: 'Travels' } }
-            },
-            {
-              name: 'visaApplication',
-              args: [
-                {
-                  name: 'where',
-                  type: {
-                    kind: 'INPUT_OBJECT',
-                    name: 'VisaApplicationsArgument'
-                  }
-                },
-                {
-                  name: 'orderBy',
-                  type: {
-                    kind: 'INPUT_OBJECT',
-                    name: 'VisaApplicationsOrderBy'
-                  }
-                }
-              ],
-              type: { ofType: { name: 'VisaApplications' } }
-            }
-          ]
-        }
-      }
-    });
-    const wrapper = mount(
-      <Router keyLength={0}>
-        <DomainExplorerLandingPage {...props} />
-      </Router>
-    );
-    expect(GraphQL.useGetQueryFieldsQuery).toHaveBeenCalled();
-    expect(wrapper).toMatchSnapshot();
+    expect(wrapper.find(DomainExplorerLandingPage)).toMatchSnapshot();
   });
 });
