+++ packages/management-console/src/components/Templates/DomainExplorerPage/tests/DomainExplorerPage.test.tsx
@@ -1,37 +1,12 @@
 import React from 'react';
-import { mount } from 'enzyme';
+import { shallow, mount } from 'enzyme';
 import { BrowserRouter } from 'react-router-dom';
-import { GraphQL } from '@kogito-apps/common';
-import DomainExplorerPage from '../DomainExplorerPage';
 import { MockedProvider } from '@apollo/react-testing';
-import gql from 'graphql-tag';
-
-jest.mock('react-apollo');
-const GET_QUERY_FIELDS = gql`
-  query getQueryFields {
-    __type(name: "Query") {
-      name
-      fields {
-        name
-        args {
-          name
-          type {
-            kind
-            name
-          }
-        }
-        type {
-          ofType {
-            name
-          }
-        }
-      }
-    }
-  }
-`;
+import DomainExplorerPage from '../DomainExplorerPage';
 
 const props = {
-  domains: ['Travels', 'VisaApplications']
+  domains: ['Travels', 'VisaApplications'],
+  loadingState: false
 };
 
 const routeComponentPropsMock = {
@@ -50,7 +25,10 @@ const routeComponentPropsMock = {
 };
 const routeComponentPropsMock2 = {
   history: {} as any,
-  location: { pathname: '/DomainExplorer/Travels', state: {} } as any,
+  location: {
+    pathname: '/DomainExplorer/Travels',
+    state: {}
+  } as any,
   match: {
     params: {
       domainName: 'Travels'
@@ -67,74 +45,18 @@ const props2 = {
     params: {
       domainName: 'Travels'
     }
-  }
+  },
+  loadingState: false
 };
 
-jest.mock('@kogito-apps/common/src/graphql/types');
 describe('Domain Explorer Dashboard component', () => {
   it('Snapshot test', () => {
-    // @ts-ignore
-    GraphQL.useGetColumnPickerAttributesQuery.mockReturnValue({
-      loading: false,
-      data: {
-        __type: {
-          fields: [
-            {
-              name: 'flight',
-              type: {
-                name: 'Flight',
-                kind: 'OBJECT',
-                fields: [
-                  {
-                    name: 'arrival',
-                    type: {
-                      name: 'String',
-                      kind: 'SCALAR'
-                    }
-                  }
-                ]
-              }
-            },
-            {
-              name: 'id',
-              type: {
-                name: 'String',
-                kind: 'SCALAR',
-                fields: null
-              }
-            }
-          ]
-        }
-      }
-    });
-    // @ts-ignore
-    GraphQL.useGetQueryFieldsQuery.mockReturnValue({
-      loading: false,
-      data: {
-        __type: {
-          fields: [
-            {
-              name: 'Travels'
-            },
-            {
-              name: 'visaApplication'
-            },
-            {
-              name: 'Jobs'
-            }
-          ]
-        }
-      }
-    });
-    // @ts-ignore
-    GraphQL.useGetQueryTypesQuery.mockReturnValue({
-      loading: false,
-      data: {}
-    });
     const wrapper = mount(
-      <BrowserRouter>
-        <DomainExplorerPage {...props} {...routeComponentPropsMock} />
-      </BrowserRouter>
+      <MockedProvider mocks={[]} addTypename={false}>
+        <BrowserRouter>
+          <DomainExplorerPage {...props} {...routeComponentPropsMock} />
+        </BrowserRouter>
+      </MockedProvider>
     );
 
     wrapper.update();
@@ -142,12 +64,6 @@ describe('Domain Explorer Dashboard component', () => {
     expect(wrapper).toMatchSnapshot();
   });
   it('Check error response for getQueryFields query', async () => {
-    // @ts-ignore
-    GraphQL.useGetQueryFieldsQuery.mockReturnValue({
-      loading: false,
-      data: null,
-      error: {}
-    });
     const wrapper = mount(
       <BrowserRouter>
         <MockedProvider mocks={[]} addTypename={false}>
@@ -160,92 +76,9 @@ describe('Domain Explorer Dashboard component', () => {
     expect(wrapper.find(DomainExplorerPage)).toMatchSnapshot();
   });
   it('Mock query testing', async () => {
-    // @ts-ignore
-    GraphQL.useGetQueryFieldsQuery.mockReturnValue({
-      loading: false,
-      data: {
-        __type: {
-          fields: [
-            {
-              name: 'Travels'
-            },
-            {
-              name: 'visaApplication'
-            },
-            {
-              name: 'Jobs'
-            }
-          ]
-        }
-      }
-    });
-    // @ts-ignore
-    GraphQL.useGetColumnPickerAttributesQuery.mockReturnValue({
-      loading: false,
-      data: {
-        __type: {
-          fields: [
-            {
-              name: 'flight',
-              type: {
-                name: 'Flight',
-                kind: 'OBJECT',
-                fields: [
-                  {
-                    name: 'arrival',
-                    type: {
-                      name: 'String',
-                      kind: 'SCALAR'
-                    }
-                  }
-                ]
-              }
-            },
-            {
-              name: 'id',
-              type: {
-                name: 'String',
-                kind: 'SCALAR',
-                fields: null
-              }
-            }
-          ]
-        }
-      }
-    });
-    // @ts-ignore
-    GraphQL.useGetQueryTypesQuery.mockReturnValue({
-      loading: false,
-      data: {}
-    });
-    const mocks = [
-      {
-        request: {
-          query: GET_QUERY_FIELDS
-        },
-        result: {
-          loading: false,
-          data: {
-            __type: {
-              fields: [
-                {
-                  name: 'Travels'
-                },
-                {
-                  name: 'visaApplication'
-                },
-                {
-                  name: 'Jobs'
-                }
-              ]
-            }
-          }
-        }
-      }
-    ];
     const wrapper = mount(
       <BrowserRouter>
-        <MockedProvider mocks={mocks} addTypename={false}>
+        <MockedProvider mocks={[]} addTypename={false}>
           <DomainExplorerPage {...props} {...routeComponentPropsMock} />
         </MockedProvider>
       </BrowserRouter>
@@ -253,19 +86,9 @@ describe('Domain Explorer Dashboard component', () => {
     wrapper.update();
     wrapper.setProps({});
     expect(wrapper.find(DomainExplorerPage)).toMatchSnapshot();
-    expect(GraphQL.useGetQueryFieldsQuery).toHaveBeenCalled();
-    expect(GraphQL.useGetQueryTypesQuery).toHaveBeenCalled();
-    expect(GraphQL.useGetColumnPickerAttributesQuery).toBeCalledWith({
-      variables: { columnPickerType: 'Travels' }
-    });
   });
   it('Check error response for getPicker query', () => {
-    // @ts-ignore
-    GraphQL.useGetColumnPickerAttributesQuery.mockReturnValue({
-      loading: false,
-      error: {}
-    });
-    const wrapper = mount(
+    const wrapper = shallow(
       <BrowserRouter>
         <DomainExplorerPage {...props} {...routeComponentPropsMock} />
       </BrowserRouter>
@@ -275,13 +98,7 @@ describe('Domain Explorer Dashboard component', () => {
     expect(wrapper).toMatchSnapshot();
   });
   it('Check error response for getQueryTypes', () => {
-    // @ts-ignore
-    GraphQL.useGetQueryTypesQuery.mockReturnValue({
-      loading: false,
-      data: null,
-      error: {}
-    });
-    const wrapper = mount(
+    const wrapper = shallow(
       <BrowserRouter>
         <DomainExplorerPage {...props} {...routeComponentPropsMock} />
       </BrowserRouter>
@@ -293,7 +110,9 @@ describe('Domain Explorer Dashboard component', () => {
   it('check assertions on rememberedParams', () => {
     const wrapper = mount(
       <BrowserRouter>
-        <DomainExplorerPage {...props2} {...routeComponentPropsMock2} />
+        <MockedProvider mocks={[]} addTypename={false}>
+          <DomainExplorerPage {...props2} {...routeComponentPropsMock2} />
+        </MockedProvider>
       </BrowserRouter>
     );
     wrapper.update();
