+++ packages/management-console/src/components/Templates/ProcessDetailsPage/ProcessDetailsPage.tsx
@@ -43,9 +43,8 @@ interface MatchProps {
   instanceID: string;
 }
 
-const ProcessDetailsPage: React.FC<
-  RouteComponentProps<MatchProps, {}, {}> & InjectedOuiaProps
-> = ({ ouiaContext, ...props }) => {
+const ProcessDetailsPage: React.FC<RouteComponentProps<MatchProps, {}, {}> &
+  InjectedOuiaProps> = ({ ouiaContext, ...props }) => {
   const id = props.match.params.instanceID;
   const [isSkipModalOpen, setIsSkipModalOpen] = useState<boolean>(false);
   const [isRetryModalOpen, setIsRetryModalOpen] = useState<boolean>(false);
