+++ packages/management-console/src/components/Templates/ProcessListPage/ProcessListPage.tsx
@@ -13,7 +13,8 @@ import {
   KogitoEmptyState,
   KogitoEmptyStateType,
   ouiaPageTypeAndObjectId,
-  ServerErrors
+  ServerErrors,
+  LoadMore
 } from '@kogito-apps/common';
 import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
@@ -21,7 +22,6 @@ import PageTitle from '../../Molecules/PageTitle/PageTitle';
 import ProcessListToolbar from '../../Molecules/ProcessListToolbar/ProcessListToolbar';
 import './ProcessListPage.css';
 import ProcessListTable from '../../Organisms/ProcessListTable/ProcessListTable';
-import LoadMore from '../../Atoms/LoadMore/LoadMore';
 import ProcessListModal from '../../Atoms/ProcessListModal/ProcessListModal';
 import axios from 'axios';
 import { InfoCircleIcon } from '@patternfly/react-icons';
